requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0,<4.43.0
huggingface_hub
Pillow>=10.0.0
pandas>=2.0.0
open-clip-torch
plotly>=5.18.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk
tqdm>=4.65.0
